steamsite/requirements.txt
certifi==2018.1.18
chardet==3.0.4
Django==2.0.3
idna==2.6
mysqlclient==1.3.12
progressbar2==3.36.0
python-utils==2.3.0
pytz==2018.3
requests==2.18.4
six==1.11.0
SteamAPI==0.1.2
urllib3==1.22
